git-svn-id: https://cbilsvn.pmacs.upenn.edu/svn/gus/FgpUtil/branches/gus4@14094 4235af6a-31f9-0310-9ad3-ecf6348f2534

+++ Util/src/test/java/org/gusdb/fgputil/db/platform/DBPlatformTest.java
@@ -17,7 +17,7 @@ public class DBPlatformTest {
     @Override public short getMaxActive() { return 20; }
     @Override public short getMaxIdle() { return 20; }
     @Override public short getMinIdle() { return 5; }
-    @Override public short getMaxWait() { return 2000; }
+    @Override public long getMaxWait() { return 2000; }
     @Override public boolean isShowConnections() { return true; }
     @Override public long getShowConnectionsInterval() { return 5; }
     @Override public long getShowConnectionsDuration() { return 30; }
